requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
aiodns==3.0.0
    # via aiohttp
aiohttp[speedups]==3.7.4.post0
    # via
    #   aiohttp-client-cache
    #   disnake
aiohttp-client-cache==0.5.2
    # via -r requirements.in
aiosqlite==0.17.0
    # via -r requirements.in
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via
    #   aiohttp
    #   aiohttp-client-cache
autopep8==1.6.0
    # via -r requirements.in
brotlipy==0.7.0
    # via aiohttp
cchardet==2.1.7
    # via aiohttp
cffi==1.15.0
    # via
    #   brotlipy
    #   pycares
chardet==4.0.0
    # via aiohttp
click==8.0.3
    # via flask
disnake==2.2.2
    # via -r requirements.in
flask==2.0.2
    # via -r requirements.in
idna==3.3
    # via yarl
itsdangerous==2.0.1
    # via
    #   aiohttp-client-cache
    #   flask
jinja2==3.0.3
    # via flask
markupsafe==2.0.1
    # via jinja2
multidict==5.2.0
    # via
    #   aiohttp
    #   yarl
pycares==4.1.2
    # via aiodns
pycodestyle==2.8.0
    # via autopep8
pycparser==2.21
    # via cffi
python-dotenv==0.19.2
    # via -r requirements.in
python-forge==18.6.0
    # via aiohttp-client-cache
six==1.16.0
    # via url-normalize
timeago==1.0.15
    # via -r requirements.in
tinydb==4.5.2
    # via -r requirements.in
toml==0.10.2
    # via autopep8
typing-extensions==4.0.1
    # via
    #   aiohttp
    #   aiosqlite
url-normalize==1.4.3
    # via aiohttp-client-cache
werkzeug==2.0.2
    # via flask
yarl==1.7.2
    # via aiohttp
